+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/MessageComposerViewModel.swift
@@ -533,11 +533,14 @@ open class MessageComposerViewModel: ObservableObject {
             log.error("Failed to write image to local temporary file")
             return
         }
+        let scale = image.scale
         let addedImage = AddedAsset(
             image: image,
             id: UUID().uuidString,
             url: imageURL,
-            type: .image
+            type: .image,
+            originalWidth: Double(image.size.width * scale),
+            originalHeight: Double(image.size.height * scale)
         )
         addedAssets.append(addedImage)
     }
@@ -1086,7 +1089,11 @@ class MessageAttachmentsConverter {
                 id: videoAttachment.id.rawValue,
                 url: localUrl,
                 type: .video,
-                extraData: videoAttachment.extraData ?? [:]
+                extraData: videoAttachment.extraData ?? [:],
+                originalWidth: videoAttachment.originalWidth,
+                originalHeight: videoAttachment.originalHeight,
+                duration: videoAttachment.duration,
+                payload: nil
             )
         }
 
@@ -1096,6 +1103,9 @@ class MessageAttachmentsConverter {
             url: videoAttachment.videoURL,
             type: .video,
             extraData: videoAttachment.extraData ?? [:],
+            originalWidth: videoAttachment.originalWidth,
+            originalHeight: videoAttachment.originalHeight,
+            duration: videoAttachment.duration,
             payload: videoAttachment.payload
         )
     }
@@ -1116,7 +1126,10 @@ class MessageAttachmentsConverter {
                 id: imageAttachment.id.rawValue,
                 url: localFileUrl,
                 type: .image,
-                extraData: imageAttachment.extraData ?? [:]
+                extraData: imageAttachment.extraData ?? [:],
+                originalWidth: imageAttachment.originalWidth,
+                originalHeight: imageAttachment.originalHeight,
+                payload: nil
             )
             completion(imageAsset)
             return
@@ -1135,6 +1148,8 @@ class MessageAttachmentsConverter {
                     url: imageAttachment.imageURL,
                     type: .image,
                     extraData: imageAttachment.extraData ?? [:],
+                    originalWidth: imageAttachment.originalWidth,
+                    originalHeight: imageAttachment.originalHeight,
                     payload: imageAttachment.payload
                 )
                 completion(imageAsset)